requirements.txt
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
numpy
opencv-python
Pillow
torch>=1.7
torchvision
tqdm
loguru
onnxscript
pnnx
ncnn
onnxsim==0.4.36